refactor(task): separate deletion services behind domain interfaces
