+++ chat2db-community-server/chat2db-community-domain/chat2db-community-domain-core/src/test/java/ai/chat2db/community/domain/core/impl/task/TaskExecutorRegistryTest.java
@@ -15,6 +15,7 @@ import ai.chat2db.community.domain.api.model.task.TaskTargetSnapshot;
 import ai.chat2db.community.domain.api.model.task.TaskType;
 import ai.chat2db.community.domain.api.model.task.extension.TaskOperation;
 import ai.chat2db.community.domain.api.model.task.extension.TaskSubmissionContext;
+import ai.chat2db.community.domain.api.service.task.ArtifactService;
 import ai.chat2db.community.domain.api.service.task.TaskExecutionContext;
 import ai.chat2db.community.domain.api.service.task.TaskExecutor;
 import ai.chat2db.community.domain.api.service.task.TaskStorage;
@@ -117,9 +118,9 @@ class TaskExecutorRegistryTest {
                     draftReference.set(draft);
                     context.write("value");
                 });
-        ArtifactService failingArtifactService = new ArtifactService() {
+        ArtifactService failingArtifactService = new ArtifactServiceImpl() {
             @Override
-            String publish(ArtifactDraft ignored) {
+            public String publish(ArtifactDraft ignored) {
                 throw new IllegalStateException("Publish failed");
             }
         };
@@ -146,9 +147,9 @@ class TaskExecutorRegistryTest {
                 exportExecutor(TaskType.QUERY_RESULT_EXPORT.name(),
                         (spec, context) -> {}),
                 importExecutor(TaskType.DATA_FILE_IMPORT.name())));
-        taskManager = new LocalTaskManager(storage, registry, new ArtifactService(),
+        taskManager = new LocalTaskManager(storage, registry, new ArtifactServiceImpl(),
                 new ConnectionContextConverter(), emptyExtensionManager(), 1, 1);
-        return new TaskServiceImpl(storage, taskManager, new ArtifactService());
+        return new TaskServiceImpl(storage, taskManager, new TaskDeletionServiceImpl(storage, new ArtifactServiceImpl()));
     }
 
     private TaskExtensionManager emptyExtensionManager() {
