+++ chat2db-community-server/chat2db-community-domain/chat2db-community-domain-core/src/test/java/ai/chat2db/community/domain/core/impl/task/TaskServiceImplTest.java
@@ -49,7 +49,7 @@ class TaskServiceImplTest {
                 task(1L, 10L, 100L, ownedArtifact),
                 task(2L, 20L, 100L, otherUserArtifact),
                 task(3L, 10L, 200L, otherOrganizationArtifact)));
-        TaskServiceImpl service = new TaskServiceImpl(storage, null, new ArtifactService());
+        TaskServiceImpl service = new TaskServiceImpl(storage, null, new TaskDeletionServiceImpl(storage, new ArtifactServiceImpl()));
         ContextUtils.setContext(Context.builder()
                 .loginUser(LoginUser.builder().id(10L).build())
                 .organizationId(100L)
